+++ internal/client/api/secret.go
@@ -13,7 +13,7 @@ func (c *Client) CreateSecret(ctx context.Context, req models.SecretRequest, tok
 	resp := &models.SecretResponse{}
 	err := c.doRequest(ctx, "POST", "/api/user/secrets", &token, req, resp)
 	if err != nil {
-		return nil, fmt.Errorf("create secret failed: %w", err)
+		return nil, fmt.Errorf("Create secret failed: %w\n", err)
 	}
 	return resp, nil
 }
@@ -24,30 +24,31 @@ func (c *Client) GetSecret(ctx context.Context, id uuid.UUID, token string) (*mo
 	resp := &models.SecretResponse{}
 	err := c.doRequest(ctx, "GET", "/api/user/secrets/"+id.String(), &token, nil, resp)
 	if err != nil {
-		return nil, fmt.Errorf("create secret failed: %w", err)
-	}
-	return resp, nil
-}
-
-// UpdateSecret performs update secret request
-func (c *Client) UpdateSecret(ctx context.Context, id uuid.UUID, token string) (*models.SecretResponse, error) {
-	// PUT /api/user/secrets/{id}
-	resp := &models.SecretResponse{}
-	err := c.doRequest(ctx, "PUT", "/api/user/secrets/"+id.String(), &token, nil, resp)
-	if err != nil {
-		return nil, fmt.Errorf("create secret failed: %w", err)
+		return nil, fmt.Errorf("Get secret failed: %w\n", err)
 	}
 	return resp, nil
 }
 
 // DeleteSecret performs delete secret request
-func (c *Client) DeleteSecret(ctx context.Context, id uuid.UUID, token string) error {
+func (c *Client) DeleteSecret(ctx context.Context, id uuid.UUID, token string) (*models.SecretResponse, error) {
 	// DELETE /api/user/secrets/{id}
+	resp := &models.SecretResponse{}
 	err := c.doRequest(ctx, "DELETE", "/api/user/secrets/"+id.String(), &token, nil, nil)
 	if err != nil {
-		return fmt.Errorf("create secret failed: %w", err)
+		return nil, fmt.Errorf("Delete secret failed: %w\n", err)
 	}
-	return nil
+	return resp, nil
+}
+
+// UpdateSecret performs update secret request
+func (c *Client) UpdateSecret(ctx context.Context, id uuid.UUID, req models.SecretRequest, token string) (*models.SecretResponse, error) {
+	// PUT /api/user/secrets/{id}
+	resp := &models.SecretResponse{}
+	err := c.doRequest(ctx, "PUT", "/api/user/secrets/"+id.String(), &token, req, resp)
+	if err != nil {
+		return nil, fmt.Errorf("Update secret failed: %w\n", err)
+	}
+	return resp, nil
 }
 
 // Sync performs syncing data
