+++ internal/client/cli/secret.go
@@ -1,12 +1,22 @@
 package cli
 
 import (
-	"errors"
+	"context"
+	"encoding/json"
 	"fmt"
+	"github.com/google/uuid"
 	"github.com/rookgm/gophkeeper/internal/models"
 	"github.com/spf13/cobra"
-	"os"
 )
+
+type SecretService interface {
+	AddCredentials(ctx context.Context, req models.Credentials, masterPassword string) (*models.Credentials, error)
+	AddText(ctx context.Context, req models.TextData, masterPassword string) (*models.TextData, error)
+	AddBinary(ctx context.Context, req models.BinaryData, masterPassword string) (*models.BinaryData, error)
+	AddBankCard(ctx context.Context, req models.BankCard, masterPassword string) (*models.BankCard, error)
+	GetSecret(ctx context.Context, id uuid.UUID, masterPassword string) (*models.SecretResponse, error)
+	DeleteSecret(ctx context.Context, id uuid.UUID, masterPassword string) (*models.SecretResponse, error)
+}
 
 type secretCmd struct {
 	secretSvc      SecretService
@@ -65,244 +75,34 @@ func (c *secretCmd) auth(cmd *cobra.Command, args []string) error {
 	return nil
 }
 
-// runAddCredentials adds credentials data
-//
-// command: secret add credentials
-//
-// flags
-// -n "name"
-// -l "login"
-// -e "note"
-func (c *secretCmd) runAddCredentials(cmd *cobra.Command, args []string) error {
-	name, _ := cmd.Flags().GetString("name")
-	if name == "" {
-		fmt.Print("Name: ")
-		fmt.Scanln(&name)
-	}
-
-	login, _ := cmd.Flags().GetString("login")
-	if login == "" {
-		fmt.Print("Login: ")
-		fmt.Scanln(&login)
-	}
-
-	note, _ := cmd.Flags().GetString("note")
-	if note == "" {
-		fmt.Print("Note: ")
-		fmt.Scanln(&note)
-	}
-
-	// read password for login
-	pwd, err := readPassword("password for " + login + ":")
-	if err != nil {
-		return fmt.Errorf("Error reading password: %v\n", err)
-	}
-
-	req := models.Credentials{
-		Name:     name,
-		Note:     note,
-		Login:    login,
-		Password: pwd,
-	}
-
-	resp, err := c.secretSvc.AddCredentials(cmd.Context(), req, c.masterPassword)
-	if err != nil {
-		return fmt.Errorf("Error adding credentials: %v\n", err)
-	}
-
-	fmt.Printf("Successfully added credentials, ID: %s\n", resp.ID)
-
-	return nil
-}
-
-// runAddTextCmd adds text data
-//
-// command: secret add text
-//
-// flags
-// -n "name"
-// -c "content"
-// -e "note"
-func (c *secretCmd) runAddTextCmd(cmd *cobra.Command, args []string) error {
-	name, _ := cmd.Flags().GetString("name")
-	if name == "" {
-		fmt.Print("Name: ")
-		fmt.Scanln(&name)
-	}
-
-	content, _ := cmd.Flags().GetString("content")
-	if content == "" {
-		fmt.Print("Content: ")
-		fmt.Scan(&content)
-	}
-
-	note, _ := cmd.Flags().GetString("note")
-	if note == "" {
-		fmt.Print("Note: ")
-		fmt.Scanln(&note)
-	}
-
-	req := models.TextData{
-		Name:    name,
-		Note:    note,
-		Content: content,
-	}
-
-	res, err := c.secretSvc.AddText(cmd.Context(), req, c.masterPassword)
-	if err != nil {
-		return fmt.Errorf("Error adding text: %v\n", err)
-	}
-
-	fmt.Printf("Successfully added text, ID: %s\n", res.ID)
-
-	return nil
-}
-
-// runAddBinaryCmd adds binary data
-//
-// command: secret add binary
-//
-// flags
-// -n "name"
-// -p "file_name"
-// -e "note"
-func (c *secretCmd) runAddBinaryCmd(cmd *cobra.Command, args []string) error {
-	name, _ := cmd.Flags().GetString("name")
-	if name == "" {
-		fmt.Print("Name: ")
-		fmt.Scanln(&name)
-	}
-
-	fileName, _ := cmd.Flags().GetString("filename")
-	if fileName == "" {
-		fmt.Print("Filename: ")
-		fmt.Scanln(&fileName)
-	}
-
-	note, _ := cmd.Flags().GetString("note")
-	if note == "" {
-		fmt.Print("Note: ")
-		fmt.Scanln(&note)
-	}
-
-	if _, err := os.Stat(fileName); errors.Is(err, os.ErrNotExist) {
-		return fmt.Errorf("File %s does not exist\n", fileName)
-	}
-
-	req := models.BinaryData{
-		Name:     name,
-		FileName: fileName,
-		Note:     note,
-		Data:     nil,
-	}
-
-	res, err := c.secretSvc.AddBinary(cmd.Context(), req, c.masterPassword)
-	if err != nil {
-		return fmt.Errorf("Error adding binary: %v\n", err)
-	}
-
-	fmt.Printf("Successfully added binary, ID: %s\n", res.ID)
-
-	return nil
-}
-
-// runAddBankCardCmd adds bank card data
-//
-// command: secret add card
-//
-// flags:
-// -n "name"
-// -e "note"
-func (c *secretCmd) runAddBankCardCmd(cmd *cobra.Command, args []string) error {
-	name, _ := cmd.Flags().GetString("name")
-	if name == "" {
-		fmt.Print("Name: ")
-		fmt.Scanln(&name)
-	}
-
-	note, _ := cmd.Flags().GetString("note")
-	if note == "" {
-		fmt.Print("Note: ")
-		fmt.Scanln(&note)
-	}
-
-	number, _ := cmd.Flags().GetString("number")
-	if number == "" {
-		fmt.Print("Card number: ")
-		fmt.Scanln(&number)
-	}
-
-	expmonth, _ := cmd.Flags().GetString("expmonth")
-	if expmonth == "" {
-		fmt.Print("Card expiration month: ")
-		fmt.Scanln(&expmonth)
-	}
-
-	expyear, _ := cmd.Flags().GetString("expyear")
-	if expyear == "" {
-		fmt.Print("Card expiration year: ")
-		fmt.Scanln(&expyear)
-	}
-
-	holdername, _ := cmd.Flags().GetString("holdername")
-	if holdername == "" {
-		fmt.Print("Card holder name: ")
-		fmt.Scanln(&holdername)
-	}
-
-	address, _ := cmd.Flags().GetString("address")
-	if address == "" {
-		fmt.Print("Cardholder's billing address: ")
-		fmt.Scanln(&address)
-	}
-
-	cardType, _ := cmd.Flags().GetString("type")
-	if cardType == "" {
-		fmt.Print("Card type: ")
-		fmt.Scanln(&cardType)
-	}
-
-	issue, _ := cmd.Flags().GetString("issue")
-	if issue == "" {
-		fmt.Print("Issue name: ")
-		fmt.Scanln(&issue)
-	}
-
-	// read ccv
-	cvv, err := readPassword("CCV: ")
-	if err != nil {
-		return fmt.Errorf("Error reading CVV: %v\n", err)
-	}
-
-	req := models.BankCard{
-		Name:            name,
-		Note:            note,
-		CardNumber:      number,
-		ExpirationMonth: expmonth,
-		ExpirationYear:  expyear,
-		CardHolderName:  holdername,
-		Cvv:             cvv,
-		BillingAddress:  address,
-		CardType:        cardType,
-		IssuingBank:     issue,
-	}
-
-	resp, err := c.secretSvc.AddBankCard(cmd.Context(), req, c.masterPassword)
-	if err != nil {
-		return fmt.Errorf("Error adding bank card: %v\n", err)
-	}
-
-	fmt.Printf("Successfully added bank card, ID: %s\n", resp.ID)
-
-	return nil
-}
-
 // runSecretGetCmd gets secret info
 //
 // command: secret get <secret_id>
 func (c *secretCmd) runSecretGetCmd(cmd *cobra.Command, args []string) error {
-	id := args[0]
-	fmt.Println(id)
+	id, err := uuid.Parse(args[0])
+	if err != nil {
+		return fmt.Errorf("Error parsing secret id: %v\n", err)
+	}
+	sec, err := c.secretSvc.GetSecret(cmd.Context(), id, c.masterPassword)
+	if err != nil {
+		return fmt.Errorf("Error getting secret: %v\n", err)
+	}
+
+	fmt.Printf("ID: %s\n", sec.ID)
+	fmt.Printf("Type: %s\n", sec.Type.String())
+	fmt.Printf("Name: %s\n", sec.Name)
+	if sec.Note != "" {
+		fmt.Printf("Note: %s\n", sec.Note)
+	}
+	fmt.Printf("Created: %s\n", sec.CreatedAt)
+	fmt.Printf("Updated: %s\n", sec.UpdatedAt)
+
+	data, err := json.MarshalIndent(sec.Data, "", " ")
+	if err != nil {
+		return fmt.Errorf("Error marshalling secret data: %v\n", err)
+	}
+	fmt.Printf("Data: %s\n", string(data))
+
 	return nil
 }
 
@@ -310,8 +110,10 @@ func (c *secretCmd) runSecretGetCmd(cmd *cobra.Command, args []string) error {
 //
 // command: secret delete <secret_id>
 func (c *secretCmd) runSecretDeleteCmd(cmd *cobra.Command, args []string) error {
-	id := args[0]
-	fmt.Println(id)
+	id, err := uuid.Parse(args[0])
+	if err != nil {
+		return fmt.Errorf("Error parsing secret id: %v\n", err)
+	}
 
 	var resp string
 	fmt.Printf("Are you sure you want to delete the secret %s? (y/N)", id)
@@ -319,6 +121,9 @@ func (c *secretCmd) runSecretDeleteCmd(cmd *cobra.Command, args []string) error 
 
 	if resp != "y" && resp != "Y" {
 		return nil
+	}
+	if _, err := c.secretSvc.DeleteSecret(cmd.Context(), id, c.masterPassword); err != nil {
+		return fmt.Errorf("Error deleting secret: %v\n", err)
 	}
 
 	return nil
