+++ internal/client/service/secret.go
@@ -11,8 +11,8 @@ import (
 type APIClientSecreter interface {
 	CreateSecret(ctx context.Context, req models.SecretRequest, token string) (*models.SecretResponse, error)
 	GetSecret(ctx context.Context, id uuid.UUID, token string) (*models.SecretResponse, error)
-	UpdateSecret(ctx context.Context, id uuid.UUID, token string) (*models.SecretResponse, error)
-	DeleteSecret(ctx context.Context, id uuid.UUID, token string) error
+	DeleteSecret(ctx context.Context, id uuid.UUID, token string) (*models.SecretResponse, error)
+	UpdateSecret(ctx context.Context, id uuid.UUID, req models.SecretRequest, token string) (*models.SecretResponse, error)
 }
 
 type AESGSMEncryptor interface {
@@ -33,14 +33,14 @@ func NewSecretService(apiClient APIClientSecreter, encryptor AESGSMEncryptor, to
 }
 
 // AddCredentials adds secret credential data
-func (c *SecretService) AddCredentials(ctx context.Context, req models.Credentials, masterPassword string) (*models.Credentials, error) {
+func (s *SecretService) AddCredentials(ctx context.Context, req models.Credentials, masterPassword string) (*models.Credentials, error) {
 	// marshaling credentials
 	credData, err := json.Marshal(req)
 	if err != nil {
 		return nil, fmt.Errorf("Error marshalling credentials: %v\n", err)
 	}
 	// encrypt data with master password
-	credEnc, err := c.encryptor.EncryptPwd(credData, masterPassword)
+	credEnc, err := s.encryptor.EncryptPwd(credData, masterPassword)
 	if err != nil {
 		return nil, fmt.Errorf("Error encrypting credentials: %v\n", err)
 	}
@@ -52,12 +52,12 @@ func (c *SecretService) AddCredentials(ctx context.Context, req models.Credentia
 		Data: credEnc,
 	}
 	// load token
-	token, err := c.tokener.Load()
+	token, err := s.tokener.Load()
 	if err != nil {
 		return nil, fmt.Errorf("Error loading token: %v\n", err)
 	}
 	// sending a request to create a secret
-	respSec, err := c.apiClient.CreateSecret(ctx, secReq, token)
+	respSec, err := s.apiClient.CreateSecret(ctx, secReq, token)
 	if err != nil {
 		return nil, fmt.Errorf("Error creating secret: %v\n", err)
 	}
@@ -70,14 +70,14 @@ func (c *SecretService) AddCredentials(ctx context.Context, req models.Credentia
 }
 
 // AddText adds secret text data
-func (c *SecretService) AddText(ctx context.Context, req models.TextData, masterPassword string) (*models.TextData, error) {
+func (s *SecretService) AddText(ctx context.Context, req models.TextData, masterPassword string) (*models.TextData, error) {
 	// marshaling text data
 	data, err := json.Marshal(req)
 	if err != nil {
 		return nil, fmt.Errorf("Error marshalling text data: %v\n", err)
 	}
 	// encrypt text data
-	textEnc, err := c.encryptor.EncryptPwd(data, masterPassword)
+	textEnc, err := s.encryptor.EncryptPwd(data, masterPassword)
 	if err != nil {
 		return nil, fmt.Errorf("Error encrypting text data: %v\n", err)
 	}
@@ -89,12 +89,12 @@ func (c *SecretService) AddText(ctx context.Context, req models.TextData, master
 		Data: textEnc,
 	}
 	// load token
-	token, err := c.tokener.Load()
+	token, err := s.tokener.Load()
 	if err != nil {
 		return nil, fmt.Errorf("Error loading token: %v\n", err)
 	}
 	// sending a request to create a secret
-	respSec, err := c.apiClient.CreateSecret(ctx, secReq, token)
+	respSec, err := s.apiClient.CreateSecret(ctx, secReq, token)
 	if err != nil {
 		return nil, fmt.Errorf("Error creating secret: %v\n", err)
 	}
@@ -106,14 +106,14 @@ func (c *SecretService) AddText(ctx context.Context, req models.TextData, master
 }
 
 // AddBinary adds secret binary data
-func (c *SecretService) AddBinary(ctx context.Context, req models.BinaryData, masterPassword string) (*models.BinaryData, error) {
+func (s *SecretService) AddBinary(ctx context.Context, req models.BinaryData, masterPassword string) (*models.BinaryData, error) {
 	// marshaling binary data
 	data, err := json.Marshal(req)
 	if err != nil {
 		return nil, fmt.Errorf("Error marshalling binary data: %v\n", err)
 	}
 	// encrypt binary data
-	binEnc, err := c.encryptor.EncryptPwd(data, masterPassword)
+	binEnc, err := s.encryptor.EncryptPwd(data, masterPassword)
 	if err != nil {
 		return nil, fmt.Errorf("Error encrypting text data: %v\n", err)
 	}
@@ -125,12 +125,12 @@ func (c *SecretService) AddBinary(ctx context.Context, req models.BinaryData, ma
 		Data: binEnc,
 	}
 	// load token
-	token, err := c.tokener.Load()
+	token, err := s.tokener.Load()
 	if err != nil {
 		return nil, fmt.Errorf("Error loading token: %v\n", err)
 	}
 	// sending a request to create a secret
-	respSec, err := c.apiClient.CreateSecret(ctx, secReq, token)
+	respSec, err := s.apiClient.CreateSecret(ctx, secReq, token)
 	if err != nil {
 		return nil, fmt.Errorf("Error creating secret: %v\n", err)
 	}
@@ -142,14 +142,14 @@ func (c *SecretService) AddBinary(ctx context.Context, req models.BinaryData, ma
 }
 
 // AddBankCard adds secret bank data
-func (c *SecretService) AddBankCard(ctx context.Context, req models.BankCard, masterPassword string) (*models.BankCard, error) {
+func (s *SecretService) AddBankCard(ctx context.Context, req models.BankCard, masterPassword string) (*models.BankCard, error) {
 	// marshaling bank card
 	data, err := json.Marshal(req)
 	if err != nil {
 		return nil, fmt.Errorf("Error marshalling bank card: %v\n", err)
 	}
 	// encrypt bank card
-	cardEnc, err := c.encryptor.EncryptPwd(data, masterPassword)
+	cardEnc, err := s.encryptor.EncryptPwd(data, masterPassword)
 	if err != nil {
 		return nil, fmt.Errorf("Error encrypting text data: %v\n", err)
 	}
@@ -161,12 +161,12 @@ func (c *SecretService) AddBankCard(ctx context.Context, req models.BankCard, ma
 		Data: cardEnc,
 	}
 	// load token
-	token, err := c.tokener.Load()
+	token, err := s.tokener.Load()
 	if err != nil {
 		return nil, fmt.Errorf("Error loading token: %v\n", err)
 	}
 	// sending a request to create a secret
-	respSec, err := c.apiClient.CreateSecret(ctx, secReq, token)
+	respSec, err := s.apiClient.CreateSecret(ctx, secReq, token)
 	if err != nil {
 		return nil, fmt.Errorf("Error creating secret: %v\n", err)
 	}
@@ -177,14 +177,96 @@ func (c *SecretService) AddBankCard(ctx context.Context, req models.BankCard, ma
 	}, nil
 }
 
-func (c *SecretService) DeleteSecret(ctx context.Context, id uuid.UUID, token string) error {
-	return nil
+// GetSecret gets secret data
+func (s *SecretService) GetSecret(ctx context.Context, id uuid.UUID, masterPassword string) (*models.SecretResponse, error) {
+	token, err := s.tokener.Load()
+	if err != nil {
+		return nil, fmt.Errorf("Error loading token: %v\n", err)
+	}
+	// get secret with encrypted data
+	respEnc, err := s.apiClient.GetSecret(ctx, id, token)
+	if err != nil {
+		return nil, fmt.Errorf("Error getting secret: %v\n", err)
+	}
+	// decrypt data
+	dec, err := s.encryptor.DecryptPwd(respEnc.Data, masterPassword)
+	if err != nil {
+		return nil, fmt.Errorf("Error decrypting secret data: %v\n", err)
+	}
+	return &models.SecretResponse{
+		ID:        respEnc.ID,
+		Name:      respEnc.Name,
+		Type:      respEnc.Type,
+		Note:      respEnc.Note,
+		Data:      dec,
+		CreatedAt: respEnc.CreatedAt,
+		UpdatedAt: respEnc.UpdatedAt,
+	}, nil
 }
 
-func (c *SecretService) GetSecret(ctx context.Context, id uuid.UUID, token string) (*models.Secret, error) {
-	return nil, nil
+// DeleteSecret deletes secret
+func (s *SecretService) DeleteSecret(ctx context.Context, id uuid.UUID, masterPassword string) (*models.SecretResponse, error) {
+	token, err := s.tokener.Load()
+	if err != nil {
+		return nil, fmt.Errorf("Error loading token: %v\n", err)
+	}
+	// get secret with encrypted data
+	respEnc, err := s.apiClient.DeleteSecret(ctx, id, token)
+	if err != nil {
+		return nil, fmt.Errorf("Error getting secret: %v\n", err)
+	}
+	// decrypt data
+	dec, err := s.encryptor.DecryptPwd(respEnc.Data, masterPassword)
+	if err != nil {
+		return nil, fmt.Errorf("Error decrypting secret data: %v\n", err)
+	}
+	return &models.SecretResponse{
+		ID:        respEnc.ID,
+		Name:      respEnc.Name,
+		Type:      respEnc.Type,
+		Note:      respEnc.Note,
+		Data:      dec,
+		CreatedAt: respEnc.CreatedAt,
+		UpdatedAt: respEnc.UpdatedAt,
+	}, nil
 }
 
-func (c *SecretService) ListSecrets(ctx context.Context, token string) ([]models.Secret, error) {
-	return nil, nil
+func (s *SecretService) UpdateSecret(ctx context.Context, id uuid.UUID, req models.SecretRequest, masterPassword string) (*models.SecretResponse, error) {
+	// encrypt data
+	dataEnc, err := s.encryptor.EncryptPwd(req.Data, masterPassword)
+	if err != nil {
+		return nil, fmt.Errorf("Error encrypting secret data: %v\n", err)
+	}
+	// load token
+	token, err := s.tokener.Load()
+	if err != nil {
+		return nil, fmt.Errorf("Error loading token: %v\n", err)
+	}
+
+	reqEnc := models.SecretRequest{
+		Name: req.Name,
+		Type: req.Type,
+		Note: req.Note,
+		Data: dataEnc,
+	}
+
+	respEnc, err := s.apiClient.UpdateSecret(ctx, id, reqEnc, token)
+	if err != nil {
+		return nil, fmt.Errorf("Error updating secret: %v\n", err)
+	}
+	dec, err := s.encryptor.DecryptPwd(respEnc.Data, masterPassword)
+	if err != nil {
+		return nil, fmt.Errorf("Error decrypting secret data: %v\n", err)
+	}
+
+	return &models.SecretResponse{
+			ID:        respEnc.ID,
+			Name:      respEnc.Name,
+			Type:      respEnc.Type,
+			Note:      respEnc.Note,
+			Data:      dec,
+			CreatedAt: respEnc.CreatedAt,
+			UpdatedAt: respEnc.UpdatedAt,
+		},
+		nil
 }
